pyproject.toml
[tool.poetry]
name = "zee5-auth"
version = "0.1.0"
description = "ZEE5 downloader — login, manifest, download"
authors = ["you@example.com"]
readme = "README.md"
packages = [{include = "zee5"}]

[tool.poetry.dependencies]
python     = "^3.11"
httpx      = "^0.27"
rich       = "^13.7"
click      = "^8.1"
pyyaml     = "^6.0"
pydantic   = "^2.7"
pydantic-settings = "^2.3"
python-dotenv     = "^1.0"
cryptography      = "^42.0"
platformdirs      = "^4.2"
pywidevine        = "^1.8"

[tool.poetry.group.dev.dependencies]
pytest       = "^8.2"
pytest-asyncio = "^0.23"
pytest-httpx   = "^0.30"
ruff = "^0.4"
mypy = "^1.10"

[tool.poetry.scripts]
zee5 = "zee5.cli:cli"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length    = 100
target-version = "py311"
